+++ packages/inngest/src/experimental.ts
@@ -1,12 +1,14 @@
 // AsyncLocalStorage
 
-export { DeferredFunction } from "./components/DeferredFunction.ts";
+export {
+  createDefer,
+  DeferredFunction,
+} from "./components/DeferredFunction.ts";
 export type { AsyncContext } from "./components/execution/als.ts";
 export { getAsyncCtx } from "./components/execution/als.ts";
 // Extended Traces (OpenTelemetry)
 export type { ExtendedTracesMiddlewareOptions } from "./components/execution/otel/middleware.ts";
 export { extendedTracesMiddleware } from "./components/execution/otel/middleware.ts";
 export { PublicInngestSpanProcessor as InngestSpanProcessor } from "./components/execution/otel/processor.ts";
-export { createDefer } from "./components/Inngest.ts";
 // Step Metadata
 export { metadataMiddleware } from "./components/InngestMetadata.ts";
